+++ pyanaconda/modules/boss/install_manager/install_manager.py
@@ -19,7 +19,7 @@
 
 from pyanaconda.core.signal import Signal
 from pyanaconda.dbus import DBus
-from pyanaconda.modules.boss.install_manager.installation_interface import InstallationNotRunning
+from pyanaconda.dbus.errors import InstallationNotRunning
 
 from pyanaconda import anaconda_logging
 log = anaconda_logging.get_dbus_module_logger(__name__)
