+++ tests/pyanaconda_tests/install_manager_test.py
@@ -19,9 +19,9 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
+from pyanaconda.dbus.errors import InstallationNotRunning
 from pyanaconda.dbus.observer import DBusObjectObserver
 from pyanaconda.modules.boss.install_manager.install_manager import InstallManager
-from pyanaconda.modules.boss.install_manager.installation_interface import InstallationNotRunning
 
 
 def _get_proxy(dbus_module_path, task_instance):
