+++ tests/pyanaconda_tests/kickstart_manager.py
@@ -20,9 +20,10 @@ import unittest
 import os
 from contextlib import contextmanager
 
+from pyanaconda.dbus.errors import SplitKickstartMissingIncludeError, \
+    SplitKickstartSectionParsingError
 from pyanaconda.dbus.observer import DBusObjectObserver
-from pyanaconda.modules.boss.kickstart_manager import KickstartManager,\
-    SplitKickstartSectionParsingError, SplitKickstartMissingIncludeError
+from pyanaconda.modules.boss.kickstart_manager import KickstartManager
 
 KICKSTART1 = """
 text
